5. SAP Generative AI Hub and HANA Cloud Vector Engine/5.1 Q&A RAG/code/docs-indexing/requirements.txt
boto3
unstructured
langchain
tiktoken
pypdf
hana-ml
hdbcli
langchain-community
generative-ai-hub-sdk
python-dotenv
tika
pandas
